requirements.txt
flask
sk-learn.externals
numpy
scipy

